+++ snooplyze/persistence/__init__.py
@@ -1,10 +1,12 @@
 from .base_persistence_engine import PersistenceEngine, PersistenceEngineType
 
 from .duckdb_persistence_engine import DuckDbPersistenceEngine
+from .sqlite_persistence_engine import SQLitePersistenceEngine
 from .postgresql_persistence_engine import PostgreSQLPersistenceEngine
 from .flatfile_persistence_engine import FlatFilePersistenceEngine
 from .mysql_persistence_engine import MySQLPersistenceEngine
 
 
 __all__ = ["PersistenceEngine", "PersistenceEngineType",
-            "DuckDbPersistenceEngine", "PostgreSQLPersistenceEngine", "FlatFilePersistenceEngine", "MySQLPersistenceEngine"]
+            "DuckDbPersistenceEngine", "SQLitePersistenceEngine", 
+            "PostgreSQLPersistenceEngine", "FlatFilePersistenceEngine", "MySQLPersistenceEngine"]