+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.9] - 2025-07-25
+
+### Changes
+
+- Add **SQLite** as a Persistence Engine
+  - All relevant classess and logic
+  - SQL scripts for creating structure
+
 ## [0.1.8] - 2025-07-25
 
 ### Changes